feat(icmp): envoyer un ping reel et confirmer la reception du pong
Le ping est desormais envoye via une socket raw ICMP et la reponse
recue depuis le reseau est deserialisee avec succes : la
serialisation/deserialisation du paquet fonctionne de bout en bout.

+++ src/icmp/packet/echo_request_packet.h
@@ -10,8 +10,8 @@
 #include "../../packet_processors/packet_processors.h"
 struct s_echo_request {
     t_header_icmp packet_header;
-    uint8_t identifier;
-    uint8_t sequence;
+    uint16_t identifier;
+    uint16_t sequence;
 };
 
 /**
